# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Webconsole API Documentation",
        "contact": {
            "name": "OMEC Project - Webconsole",
            "url": "https://github.com/omec-project/webconsole"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5000",
    "basePath": "/",
    "paths": {
        "/api/subscriber/": {
            "get": {
                "description": "Return the list of subscribers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscribers"
                ],
                "responses": {
                    "200": {
                        "description": "List of subscribers",
                        "schema": {
                            "$ref": "#/definitions/configmodels.SubsListIE"
                        }
                    },
                    "500": {
                        "description": "Error retrieving subscribers"
                    }
                }
            }
        },
        "/api/subscriber/{imsi}": {
            "get": {
                "description": "Get subscriber by IMSI (UE ID)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscribers"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "example": "imsi-208930100007487",
                        "description": "IMSI (UE ID)",
                        "name": "imsi",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subscriber"
                    },
                    "404": {
                        "description": "Subscriber not found"
                    },
                    "500": {
                        "description": "Error retrieving subscriber"
                    }
                }
            },
            "post": {
                "description": "Create subscriber by IMSI (UE ID)",
                "tags": [
                    "Subscribers"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "IMSI (UE ID)",
                        "name": "imsi",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": " ",
                        "name": "content",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/configmodels.SubsOverrideData"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Subscriber created"
                    },
                    "400": {
                        "description": "Invalid subscriber content"
                    },
                    "500": {
                        "description": "Error creating subscriber"
                    }
                }
            },
            "delete": {
                "description": "Delete an existing subscriber",
                "tags": [
                    "Subscribers"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "IMSI (UE ID)",
                        "name": "imsi",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Subscriber deleted successfully"
                    },
                    "500": {
                        "description": "Error deleting subscriber"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "configmodels.SubsListIE": {
            "type": "object",
            "properties": {
                "plmnID": {
                    "type": "string"
                },
                "ueId": {
                    "type": "string"
                }
            }
        },
        "configmodels.SubsOverrideData": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "opc": {
                    "type": "string"
                },
                "plmnID": {
                    "type": "string"
                },
                "sequenceNumber": {
                    "type": "string"
                }
            }
        },
...
    }
}